third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/source-frame-breakpoint-decorations-expected.txt

Checks that JavaScriptSourceFrame show breakpoints correctly


Running: testAddRemoveBreakpoint
Setting breakpoint
breakpoint at 2
  inline breakpoint at (2, 4)
  inline breakpoint at (2, 13) disabled
Toggle breakpoint

Running: testTwoBreakpointsResolvedInOneLine
Setting breakpoint
breakpoint at 2 conditional
  inline breakpoint at (2, 4)
  inline breakpoint at (2, 13) disabled
Toggle breakpoint

Running: testDecorationInGutter
Adding regular disabled breakpoint
breakpoint at 2 disabled
Adding conditional disabled breakpoint
breakpoint at 2 disabled conditional
Adding regular enabled breakpoint
breakpoint at 2
  inline breakpoint at (2, 4)
  inline breakpoint at (2, 13) disabled
Adding conditional enabled breakpoint
breakpoint at 2 conditional
  inline breakpoint at (2, 4)
  inline breakpoint at (2, 13) disabled
Disable breakpoints
breakpoint at 2 disabled conditional
Enable breakpoints
breakpoint at 2 conditional
  inline breakpoint at (2, 4)
  inline breakpoint at (2, 13) disabled
Remove breakpoints
